pyproject.toml
[tool.poetry]
name = "social-network-test"
version = "0.1.0"
description = ""
authors = ["Enmanuel Silva <enmanuel.silva@spotcloud.io>"]

[tool.poetry.dependencies]
python = "^3.9"
redis = "^4.3.4"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
